Hotfix/general fixes and warnings

+++ Project/Source/Components/ComponentParticleSystem.cpp
@@ -1,29 +1,23 @@
 #include "ComponentParticleSystem.h"
 
-#include "GameObject.h"
-#include "Components/ComponentTransform.h"
-#include "Components/UI/ComponentButton.h"
 #include "Application.h"
+#include "GameObject.h"
 #include "Modules/ModulePrograms.h"
 #include "Modules/ModuleCamera.h"
 #include "Modules/ModuleRender.h"
 #include "Modules/ModuleEditor.h"
-#include "Modules/ModuleResources.h"
 #include "Modules/ModuleTime.h"
 #include "Modules/ModuleUserInterface.h"
 #include "Modules/ModulePhysics.h"
-#include "Panels/PanelScene.h"
+#include "Components/ComponentTransform.h"
 #include "Resources/ResourceTexture.h"
 #include "Resources/ResourceShader.h"
-#include "FileSystem/TextureImporter.h"
 #include "FileSystem/JsonValue.h"
-#include "Utils/Logging.h"
 #include "Utils/ImGuiUtils.h"
+#include "Utils/ParticleMotionState.h"
 
 #include "Math/float3x3.h"
 #include "Math/TransformOps.h"
-#include "imgui.h"
-#include "imgui_internal.h"
 #include "imgui_color_gradient.h"
 #include "GL/glew.h"
 #include "debugdraw.h"
@@ -1127,7 +1121,7 @@ void ComponentParticleSystem::StopChildParticles() {
 }
 
 float ComponentParticleSystem::ChildParticlesInfo() {
-	float particlesInfo = particles.Count();
+	float particlesInfo = (float) particles.Count();
 	for (GameObject* currentChild : GetOwner().GetChildren()) {
 		if (currentChild->GetComponent<ComponentParticleSystem>()) {
 			particlesInfo += currentChild->GetComponent<ComponentParticleSystem>()->ChildParticlesInfo();
